third_party/blink/web_tests/platform/generic/storage/indexeddb/basics-workers-expected.txt

[Worker] Test IndexedDB's basics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/basics.js
[Worker] indexedDB.open('basics')
PASS [Worker] 'result' in request is true
[Worker] Expecting exception from request.result
PASS [Worker] Exception was thrown.
PASS [Worker] code is DOMException.INVALID_STATE_ERR
PASS [Worker] ename is 'InvalidStateError'
[Worker] Exception message: Failed to read the 'result' property from 'IDBRequest': The request has not finished.
PASS [Worker] 'error' in request is true
[Worker] Expecting exception from request.error
PASS [Worker] Exception was thrown.
PASS [Worker] code is DOMException.INVALID_STATE_ERR
PASS [Worker] ename is 'InvalidStateError'
[Worker] Exception message: Failed to read the 'error' property from 'IDBRequest': The request has not finished.
PASS [Worker] 'source' in request is true
PASS [Worker] request.source is null
PASS [Worker] 'transaction' in request is true
PASS [Worker] request.transaction is null
PASS [Worker] 'readyState' in request is true
PASS [Worker] request.readyState is "pending"
PASS [Worker] 'onsuccess' in request is true
PASS [Worker] request.onsuccess is null
PASS [Worker] 'onerror' in request is true
PASS [Worker] request.onerror is null
PASS [Worker] 'result' in event.target is true
PASS [Worker] !!event.target.result is true
PASS [Worker] 'error' in event.target is true
PASS [Worker] event.target.error is null
PASS [Worker] 'source' in event.target is true
PASS [Worker] request.source is null
PASS [Worker] 'transaction' in event.target is true
PASS [Worker] event.target.transaction is null
PASS [Worker] 'readyState' in request is true
PASS [Worker] event.target.readyState is "done"
PASS [Worker] 'onsuccess' in event.target is true
PASS [Worker] 'onerror' in event.target is true
PASS successfullyParsed is true

TEST COMPLETE
